Added Saving and Loading. Saving currently saves to (filename)2 for testing reasons.
added a weird way of starting the app, because Item.ItemList needed declared, but I didn't want it to be reliant on having a file to load. So, it is declared, loads a blanks list, and then clears itself. There's probably a better way to do this but I've not discovered it yet.